`no-std` compatibility for proof structure
