+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Refactored to/from_bytes criteria for some structs (#333)
 
+## [0.4.0] - 24-12-20
+### Added 
+- `no-std` compatibility for proof structure.
+
 ## [0.3.6] - 17-12-20
 ### Added 
 - To/From bytes impl for `PublicInput`.
