+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "dusk-plonk"
-version = "0.3.6"
+version = "0.4.0"
 authors = ["Kevaundray Wedderburn <kevtheappdev@gmail.com>",
            "Luke Pearson <luke@dusk.network>", 
            "CPerezz <carlos@dusk.network>"] 
@@ -21,29 +21,40 @@ license = "MPL-2.0"
 edition = "2018"
 
 [dependencies]
-merlin = "2.0.0"
-rand = "0.7.2"
-rand_core = { version = "0.5", default-features = false }
-# Built by default with "std", "alloc", "pairing", "groups" and "endo" features.
-dusk-bls12_381 = "0.3.0"
-itertools = "0.9.0"
-rand_chacha = "0.2"
-rayon = "1.3.0"
-anyhow = "1.0.32"
-dusk-jubjub = "0.5.0"
-thiserror = "1.0"
-serde = "1.0"
-# Dusk related deps for WASMI serde
+rand_core = "0.6"
+cfg-if = "1.0"
+dusk-bls12_381 = {git = "https://github.com/dusk-network/bls12_381", branch = "vlopes11/no-std-pairings", default-features = false, features = ["pairings"]}
+dusk-jubjub = {git = "https://github.com/dusk-network/jubjub", branch = "vlopes11/no-std-pairings", default-features = false}
+serde = {version = "1.0", default-features = false}
+merlin = {version = "2.0", default-features = false}
 canonical = {version = "0.4", optional = true}
-
-[dev-dependencies]
-rand = "0.7.0"
+itertools = {version = "0.9", optional = true}
+rayon = {version = "1.3", optional = true}
+anyhow = {version = "1.0", optional = true}
+thiserror = {version = "1.0", optional = true}
+rand = {version = "0.8", optional = true}
 
 [package.metadata.docs.rs]
 rustdoc-args = [ "--html-in-header", "katex-header.html" ]
 
 [features]
+default = ["std"]
+std = [
+    "dusk-bls12_381/default",
+    "dusk-jubjub/std",
+    "serde/default",
+    "merlin/default",
+    "itertools",
+    "rayon",
+    "anyhow",
+    "thiserror",
+    "rand"
+]
 nightly = []
 trace = []
 trace-print = ["trace"]
-canon = ["dusk-bls12_381/canon", "dusk-jubjub/canon", "canonical"]
+canon = [
+    "dusk-bls12_381/canon",
+    "dusk-jubjub/canon",
+    "canonical"
+]
